backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
numpy==1.24.3
umap-learn==0.5.4
geopandas==0.14.1
pandas==2.1.3
pyarrow==14.0.1